+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- `FromContext` now receives a context key as additional parameter.
+
 ## [0.5.0] - 2018-03-12
 ### Added
 - `FromContext` function to extract a JWT object from a context.
@@ -60,6 +64,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - This package's source code, including examples and tests.
 - Go dep files.
 
+[Unreleased]: https://github.com/gbrlsnchs/jwt/compare/v0.5.0...HEAD
 [0.5.0]: https://github.com/gbrlsnchs/jwt/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/gbrlsnchs/jwt/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/gbrlsnchs/jwt/compare/v0.2.0...v0.3.0